% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{computeLoglikelihood}
\alias{computeLoglikelihood}
\title{Serially computed MDS log likelihood and gradient}
\usage{
computeLoglikelihood(data, locations, precision, truncation = TRUE,
  gradient = FALSE)
}
\arguments{
\item{data}{A distance matrix.}

\item{locations}{Object positions in latent space.}

\item{precision}{Precision for MDS likelihood.}

\item{truncation}{Should truncation term be included? Defaults to TRUE.}

\item{gradient}{Return gradient (or log likelihood)? Defaults to FALSE.}
}
\value{
MDS log likelihood or its gradient.
}
\description{
Called by \code{MassiveMDS::test()} to compare output with parallel implementations.
Slow. Not recommended for use.
}
