% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{dmatrixnorm}
\alias{dmatrixnorm}
\title{Matrix normal log density and gradient}
\usage{
dmatrixnorm(X, Mu = NULL, U, V, Uinv, Vinv, gradient = FALSE)
}
\arguments{
\item{X}{An n by p matrix of n p-dimensional latent locations.}

\item{Mu}{The n by p mean of the matrix normal.}

\item{U}{The n by n covariance matrix.}

\item{V}{The p by p covariance matrix.}

\item{Uinv}{The inverse of U.}

\item{Vinv}{The inverse of V.}

\item{gradient}{Return gradient (or log likelihood)? Defaults to FALSE.}
}
\value{
Matrix normal log density or its gradient.
}
\description{
Default prior for latent locations.  Called by \code{MassiveMDS::hmcsampler()}.
}
