% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{engineInitial}
\alias{engineInitial}
\title{Initialize MDS engine object}
\usage{
engineInitial(data, locations, N, P, precision = 1, threads, simd,
  truncation, gpu, single)
}
\arguments{
\item{data}{N by N distance matrix.}

\item{locations}{N by P latent locations.}

\item{N}{Number of locations and size of distance matrix.}

\item{P}{Dimension of latent locations.}

\item{precision}{MDS likelihood precision.}

\item{threads}{Number of CPU cores to be used.}

\item{simd}{For CPU implementation: no SIMD (\code{0}), SSE (\code{1}) or AVX (\code{2}).}

\item{truncation}{Likelihood includes truncation term? Defaults to \code{TRUE}.}

\item{gpu}{Which GPU to use? If only 1 available, use \code{gpu=1}. Defaults to \code{0}, no GPU.}

\item{single}{Set \code{single=1} if your GPU does not accommodate doubles.}
}
\value{
MDS engine object.
}
\description{
Takes data, parameters and implementation details and returns MDS engine object. Used within
\code{MassiveMDS::hmcsampler()}.
}
