% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{getdata}
\alias{getdata}
\title{Read distance matrix data from file}
\usage{
getdata(N, locations, file = "h3_Deff")
}
\arguments{
\item{N}{Number of locations and size of distance matrix.}

\item{locations}{Latent locations with names to be matched to file entries.}

\item{file}{Text file located in \code{inst/extdata} containing distance matrix.}
}
\value{
Distance matrix.
}
\description{
For BEAST applications, reads N dimensional distance matrix with names matched to location names.
}
