% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{getGradient}
\alias{getGradient}
\title{Helper MDS log likelihood gradient function}
\usage{
getGradient(engine)
}
\arguments{
\item{engine}{An MDS engine object.}
}
\value{
MDS log likelihood gradient.
}
\description{
Takes MDS engine object and returns log likelihood gradient.
}
