% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{getLogLikelihood}
\alias{getLogLikelihood}
\title{Helper MDS log likelihood function}
\usage{
getLogLikelihood(engine)
}
\arguments{
\item{engine}{An MDS engine object.}
}
\value{
MDS log likelihood
}
\description{
Takes MDS engine object and returns log likelihood.
}
