% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassiveMDS.R
\docType{package}
\name{MassiveMDS}
\alias{MassiveMDS}
\alias{MassiveMDS-package}
\title{MassiveMDS}
\description{
MassiveMDS facilitates fast Bayesian MDS through GPU, multi-core CPU, and SIMD vectorization powered implementations of the Hamiltonian Monte Carlo algorithm.
The package may be built either as a standalone library or as an R package relying on Rcpp.
}
\author{
Marc Suchard and Andrew Holbrook
}
