% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{Potential}
\alias{Potential}
\title{Potential function and gradient for HMC}
\usage{
Potential(
  engine,
  locations,
  treeVcv,
  traitVcv,
  treePrec,
  traitPrec,
  gradient = FALSE
)
}
\arguments{
\item{engine}{MDS engine object.}

\item{locations}{N by P matrix of N P-dimensional latent locations.}

\item{treeVcv}{N by N covariance.}

\item{traitVcv}{P by P covariance.}

\item{treePrec}{N by N precision.}

\item{traitPrec}{P by P precision.}

\item{gradient}{Return gradient (instead of potential)? Defaults \code{FALSE}.}
}
\value{
Potential or its gradient.
}
\description{
Takes MDS engine object, latent locations and other model parameters. Returns
potential (proportional to log posterior) function or its gradient.
}
