% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{readbeast}
\alias{readbeast}
\title{Read BEAST file}
\usage{
readbeast(file = "large", priorRootSampleSize = 0.001)
}
\arguments{
\item{file}{Name of \code{.trees} file stored in \code{inst/extdata/}.}

\item{priorRootSampleSize}{Prior precision on root.}
}
\value{
List with initial locations, nxn and pxp covariances, and degrees of freedom on Wishart prior.
}
\description{
Reads BEAST \code{.trees} file and returns various quantities useful for Bayesian MDS with HMC.
}
