% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{setPairwiseData}
\alias{setPairwiseData}
\title{Deliver distance matrix to MDS engine object}
\usage{
setPairwiseData(engine, data)
}
\arguments{
\item{engine}{MDS engine object.}

\item{data}{Distance matrix.}
}
\value{
MDS engine object.
}
\description{
Helper function delivers distance matrix to MDS engine object.
}
