% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{setPrecision}
\alias{setPrecision}
\title{Deliver precision variable to MDS engine object}
\usage{
setPrecision(engine, precision)
}
\arguments{
\item{engine}{MDS engine object.}

\item{precision}{MDS likelihood precision.}
}
\value{
MDS engine object.
}
\description{
Helper function delivers MDS likelihood precision to MDS engine object.
}
