% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{test}
\alias{test}
\title{Compare serially and parallel-ly computed log likelihoods and gradients}
\usage{
test(locationCount = 10, threads = 0, simd = 0, gpu = 0,
  single = 0)
}
\arguments{
\item{locationCount}{Size of distance matrix or number of latent locations.}

\item{threads}{Number of CPU cores to be used.}

\item{simd}{For CPU implementation: no SIMD (\code{0}), SSE (\code{1}) or AVX (\code{2}).}

\item{gpu}{Which GPU to use? If only 1 available, use \code{gpu=1}. Defaults to \code{0}, no GPU.}

\item{single}{Set \code{single=1} if your GPU does not accommodate doubles.}
}
\value{
Returns MDS log likelihoods (should be equal) and distance between gradients (should be 0).
}
\description{
Compare outputs for serially and parallel-ly computed MDS log likelihoods and gradients across
a range of implementations for both truncated and non-truncated likelihoods. Randomly generates
distance matrix and latent locations.
}
