% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{timeTest}
\alias{timeTest}
\title{Time log likelihood and gradient calculations}
\usage{
timeTest(
  locationCount = 5000,
  maxIts = 1,
  threads = 0,
  simd = 0,
  gpu = 0,
  single = 0,
  bandwidth = NULL
)
}
\arguments{
\item{locationCount}{Size of distance matrix or number of latent locations.}

\item{maxIts}{Number of times to compute the log likelihood and gradient.}

\item{threads}{Number of CPU cores to be used.}

\item{simd}{For CPU implementation: no SIMD (\code{0}), SSE (\code{1}) or AVX (\code{2}).}

\item{gpu}{Which GPU to use? If only 1 available, use \code{gpu=1}. Defaults to \code{0}, no GPU.}

\item{single}{Set \code{single=1} if your GPU does not accommodate doubles.}
}
\value{
User, system and elapsed time. Elapsed time is most important.
}
\description{
Time log likelihood and gradient calculations for serial and parallel implementations.
}
