% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{updateLocations}
\alias{updateLocations}
\title{Deliver latent locations matrix to MDS engine object}
\usage{
updateLocations(engine, locations)
}
\arguments{
\item{engine}{MDS engine object.}

\item{locations}{N by P matrix of N P-dimensional latent locations.}
}
\value{
MDS engine object.
}
\description{
Helper function delivers latent locations matrix to MDS engine object.
}
